requirements.txt
# AetherEdge Platform Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
elasticsearch==8.11.0

# Machine Learning & AI
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.2

# Cloud SDKs
boto3==1.34.0
azure-identity==1.15.0
azure-mgmt-compute==30.3.0
google-cloud-compute==1.14.1

# Infrastructure as Code
terraform-compliance==1.3.44
ansible==8.7.0

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation-fastapi==0.42b0
jaeger-client==4.8.0

# Security
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT==2.8.0

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
psutil==5.9.6

# Configuration & Environment
python-dotenv==1.0.0
dynaconf==3.2.4
click==8.1.7

# Data Processing
marshmallow==3.20.2
jsonschema==4.20.0
pyyaml==6.0.1
toml==0.10.2

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1

# Testing (dev dependencies)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For testing FastAPI
factory-boy==3.3.0

# Code Quality (dev dependencies)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3
bandit==1.7.5

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-mermaid2-plugin==1.1.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
rich==13.7.0
typer==0.9.0

# Development Tools
pre-commit==3.6.0
gitpython==3.1.40

# Pinned versions for security
urllib3==2.1.0
certifi==2023.11.17
idna==3.6
charset-normalizer==3.3.2
